Crate cryptowise

The cryptowise crate provides functions that are used to answer the Matasano Crypto Challenges (cryptopals.com)

Modules

set1
The set1 module manages the declaration of the challenge modules for this set. This module is the qualifying set. The exercises here are to ramp developers up gradually into being able to code cryptography challenges. It's said that these are easy. For some I suppose. My implementation of these challenges is done using Rust and I'm a Java professional first. Rust is a good choice for these challenges.