Module cryptowise::set1

The set1 module manages the declaration of the challenge modules for this set. This module is the qualifying set. The exercises here are to ramp developers up gradually into being able to code cryptography challenges. It's said that these are easy. For some I suppose. My implementation of these challenges is done using Rust and I'm a Java professional first. Rust is a good choice for these challenges.

Modules

challenge1
This challenge is to convert a hexadecimal string respresention and prove a conversion to base 64 equates to a translation that's given.

challenge2
This challenge is to perform an exclusive bitwise OR operation on two hexadecimal strings.

challenge3
This challenge module requires you to devise a way to score a response that's obtained by doing an XOR operation in reverse. A plain-text string is XOR'ed with a single character key. The cryptopal website is a little unclear about how this character is used. The hex string given in the challenge is the result of the secret message and the single key (2 char byte represenation) repeated 34 times beneath or across the hex string length. The answer to this challenge is given by running the tests like so:

challenge4
This challenge module requires the reading of a sizeable file with 60 hex-character lines. One of the lines is encrypted using single-character XOR. It's our job to find that line. I will use the techniques learned in challenge-3 for an ultimate solution.

challenge5
This challenge requests that we encrypt a specified text string in plain-text using a repeating key XOR operation using "ICE" as the three bytes that will be repeated beneath the plain-text from left to right. ICE equates to 49,43, and 45 in base16(hex) respectively. The plain-text string must be first converted to a hex string and then the 494345494345... sequence XOR'ed with it to produce the encrypted hexadecimal string. The encrypted hex string is given in the cryptopals challenge.

challenge6
This challenge requires use to break a repeating-key XOR ("Vigenere"). This will be the qualifying challenge for the rest of the challenges. It will be difficult. The claim is that people know how to break this but less people can actually break it.