Module cryptowise::set1::challenge3

This challenge module requires you to devise a way to score a response that's obtained by doing an XOR operation in reverse. A plain-text string is XOR'ed with a single character key. The cryptopal website is a little unclear about how this character is used. The hex string given in the challenge is the result of the secret message and the single key (2 char byte represenation) repeated 34 times beneath or across the hex string length. The answer to this challenge is given by running the tests like so:

$ cargo test -- --nocapture

Test Notes

The syntax here is not correct for the nocapture option, however it does work to allow the print statements to come with the test output.

Examples

let tmp = "Mary had a little lamb.";
assert_eq!(5, word_count(tmp));

Functions

single_char_key
This function provides a usize primitive (integer value) for the best or most worded XOR result of this challenge. A loop is used to test each single-repeatable hex character, representing the alphabet (in decimal) between 65 and 123. Each decimal is converted to a 2 char hex representation and then repeated 34 times creating a string of similar lenght to the encrypted hex string given. Each XOR produces what could be the secret message and the function here does a word count. The XOR output with the most number of words is kept and when the loop is terminated, the function returns this best usize value.

word_count
This function returns a usize primitive with the integer count of a provided std::string argument (statically addressable string)