Function cryptowise::set1::challenge3::single_char_key

pub fn single_char_key() -> usize

This function provides a usize primitive (integer value) for the best or most worded XOR result of this challenge. A loop is used to test each single-repeatable hex character, representing the alphabet (in decimal) between 65 and 123. Each decimal is converted to a 2 char hex representation and then repeated 34 times creating a string of similar lenght to the encrypted hex string given. Each XOR produces what could be the secret message and the function here does a word count. The XOR output with the most number of words is kept and when the loop is terminated, the function returns this best usize value.

For this challenge the best result proved 7 words. To find the secret message you'll have to checkout this code and run the tests.

Examples