Function cryptowise::set1::challenge4::get_message

pub fn get_message(hex_line_60: &str)

This function is used when the process_file function has found a line with at least 3 words. If a line contains two words, it may be "UTF8-8 Error" and we don't want that line or message. The same loop is run again for 0 to 255 testing each single char key with the XOR operation and it prints the best matching plain-text message to the console (using -- --nocapture)