Module cryptowise::set1::challenge5

This challenge requests that we encrypt a specified text string in plain-text using a repeating key XOR operation using "ICE" as the three bytes that will be repeated beneath the plain-text from left to right. ICE equates to 49,43, and 45 in base16(hex) respectively. The plain-text string must be first converted to a hex string and then the 494345494345... sequence XOR'ed with it to produce the encrypted hexadecimal string. The encrypted hex string is given in the cryptopals challenge.

Examples

Functions

static_to_hex
This function takes a static string argument and converts it to hex returning String.